+++ packages/hds-js/index.ts
@@ -1,0 +1,4 @@
+export * from '../react/src/components/cookieConsent/cookieConsentController';
+export * from '../react/src/components/cookieConsent/cookieController';
+export * from '../react/src/components/cookieConsent/getContent';
+export * from '../react/src/components/login/index.vanilla-js';