+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Changed
 
 Changes that are not related to specific components
+
 - [Component] What has been changed
 
 #### Fixed
@@ -59,11 +60,12 @@ Changes that are not related to specific components
 #### Changed
 
 Changes that are not related to specific components
+
 - [Component] What has been changed
 
 #### Fixed
 
-- Fixed Wave motifs links to Helsinki Brand pages 
+- Fixed Wave motifs links to Helsinki Brand pages
 - [Header] Small fixes
 - [Notification] Fix size texts
 
@@ -80,6 +82,7 @@ Changes that are not related to specific components
 #### Changed
 
 Changes that are not related to specific components
+
 - [Component] What has been changed
 
 #### Fixed
@@ -99,11 +102,21 @@ Changes that are not related to specific components
 #### Changed
 
 Changes that are not related to specific components
+
 - [Component] What has been changed
 
 #### Fixed
 
 - [Component] What bugs/typos are fixed?
+
+### Hds-js
+
+#### Added
+
+- [cookieConsentController] The Cookie Consent controller for City of Helsinki cookies
+- [cookieController] Controller for reading and writing City of Helsinki cookies
+- [getContent] Common content for City of Helsinki cookies
+- [login] Vanilla JS version of the login system
 
 ## [3.3.0] - December, 19, 2023